⚡ Bolt: optimize `@for` loops by using unique `track` identifiers instead of object references
This commit addresses a performance bottleneck related to Angular's `@for` block where entire objects were being tracked across iterables (e.g. `track item`). This led to unintended performance regressions.

It updates multiple files to track elements by primitive keys instead:
- `track template.id`
- `track webhook.id`
- `track user.id`
- `track campaign.id`
- `track message.id`
- `track conversation.message.id`
- `track conversation.id`
- `track contact.name.formatted_name`
- `track phone.phone`
- `track email.email`
- `track url.url`

* Reduces unnecessary DOM diffing and component re-renders when data mutates or fetches via the store.
