+++ src/app/chats-sidebar/chats-sidebar.component.html
@@ -108,7 +108,7 @@
             <ul class="scrollable flex-1 overflow-y-auto" (scroll)="onScroll($event)">
                 @for (
                     conversation of conversationStore.searchConversations;
-                    track conversation.message.id
+                    track conversation.id
                 ) {
                     <li>
                         <app-conversation-preview
