+++ src/app/messages/message-media-content/message-media-content.component.ts
@@ -7,7 +7,6 @@ import { UseMedia } from "../../../core/message/model/media-data.model";
 import { MediaStoreService } from "../../../core/media/store/media-store.service";
 import { MatIconModule } from "@angular/material/icon";
 import { NGXLogger } from "ngx-logger";
-import { isTrustedUrl } from "../../../core/common/util/url-validator.util";
 
 @Component({
     selector: "app-message-media-content",
@@ -29,7 +28,7 @@ export class MessageMediaContentComponent implements OnInit {
     @Input() isSent!: boolean;
     @Output() asyncContentLoaded = new EventEmitter();
 
-    mediaSafeUrl: SafeUrl | string = "";
+    mediaSafeUrl: SafeUrl = "";
 
     async ngOnInit(): Promise<void> {
         await this.handleAutoPreview();
@@ -40,11 +39,7 @@ export class MessageMediaContentComponent implements OnInit {
         if (!this.mediaData) return;
         const url = this.mediaData?.link;
         if (url) {
-            if (isTrustedUrl(url)) {
-                this.mediaSafeUrl = this.sanitizer.bypassSecurityTrustUrl(url);
-            } else {
-                this.mediaSafeUrl = url; // Let Angular's template binding sanitize external URLs
-            }
+            this.mediaSafeUrl = this.sanitizer.bypassSecurityTrustUrl(url); // Sanitize the URL
             return;
         }
         if (!this.mediaData.id) return;
